+++ src/main/java/sfy/option/repository/ProjectRepository.java
@@ -3,6 +3,6 @@ package sfy.option.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sfy.option.model.entity.ProjectEntity;
 
-public interface ProjectRepository extends JpaRepository<ProjectEntity, Integer> {
+public interface ProjectRepository extends JpaRepository<ProjectEntity, Long> {
 
 }
